# backend/openapi.yml
openapi: 3.1.0
info:
  version: 1.0.0
  title: Testing Nirvana API
servers:
  - url: api
components:
...
  schemas:
...
    Pagination:
      type: object
      properties:
        pages:
          type: number
        total:
          type: number
      required:
        - pages
        - total
...
    UserWithPermissions:
      type: object
      properties:
        id:
          type: string
          format: uuid
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
          format: email
        phoneNumber:
          type:
            - string
            - "null"
        securityLevel:
          type: string
          enum:
            - FULL_ACCESS
            - LIMITED
            - READ_ONLY
        status:
          type: string
          enum:
            - ACTIVE
            - INACTIVE
        role:
          type: object
          properties:
            id:
              type: string
              format: uuid
            name:
              type: string
          required:
            - id
            - name
        createdAt:
          type: string
        updatedAt:
          type: string
        permissions:
          type: array
          items:
            type: string
            enum:
              - MANAGE_USERS
              - ASSIGN_ROLES
              - DEACTIVATE_USERS
              - RESET_PASSWORDS
              - CONFIGURE_TEST_SITES
              - SCHEDULE_TESTS
              - ASSIGN_ACCOMMODATIONS
              - TRACK_STUDENT_CONFIRMATIONS
              - MANAGE_EQUIPMENT
              - MANAGE_TESTS
              - ASSIGN_STUDENTS
              - TRACK_ATTENDANCE
              - SEND_REMINDERS
              - VIEW_REPORTS
              - SCAN_QR_CODES
              - REPORT_INCIDENTS
              - CONFIRM_TEST_LOCATION
              - VIEW_RESULT
              - VIEW_INVOICES
      required:
        - id
        - firstName
        - lastName
        - email
        - phoneNumber
        - securityLevel
        - status
        - role
        - createdAt
        - updatedAt
        - permissions
...
    Invoice:
      type: object
      properties:
        id:
          type: string
          format: uuid
        providerId:
          type: string
          format: uuid
        studentId:
          type: string
          format: uuid
        therapyServiceId:
          type: string
          format: uuid
        amount:
          type: number
        status:
          type: string
          enum:
            - PENDING
            - PAID
            - DECLINED
        dateIssued:
          type:
            - string
            - "null"
        createdAt:
          type: string
        updatedAt:
          type: string
        deletedAt:
          type:
            - string
            - "null"
        provider:
          type: object
          properties:
            id:
              type: string
              format: uuid
            user:
              type: object
              properties:
                id:
                  type: string
                  format: uuid
                firstName:
                  type: string
                lastName:
                  type: string
                email:
                  type: string
                  format: email
              required:
                - id
                - firstName
                - lastName
                - email
          required:
            - id
            - user
        student:
          type: object
          properties:
            id:
              type: string
              format: uuid
            firstName:
              type: string
            lastName:
              type: string
          required:
            - id
            - firstName
            - lastName
        therapyService:
          type: object
          properties:
            id:
              type: string
              format: uuid
            serviceType:
              type: string
              enum:
                - SPEECH
                - OCCUPATIONAL
                - PHYSICAL
          required:
            - id
            - serviceType
      required:
        - id
        - providerId
        - studentId
        - therapyServiceId
        - amount
        - status
        - dateIssued
        - createdAt
        - updatedAt
        - deletedAt
        - provider
        - student
        - therapyService
...
paths:
...
  /auth/login:
    post:
      description: Authenticate with email and password.
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
              required:
                - email
                - password
      responses:
        "200":
          description: User with permissions object.
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/UserWithPermissions"
                required:
                  - data
        "400": *a1
        "401": *a1
        "403": *a1
        "404": *a1
        "500": *a1
  /auth/start-password-reset:
    post:
      description: Request a password reset token that will be sent over email.
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
              required:
                - email
      responses:
        "204":
          description: The password reset email will be sent if the user exists.
        "400": *a1
        "401": *a1
        "403": *a1
        "404": *a1
        "500": *a1
  /auth/reset-password:
    post:
      description: Reset a user's password.
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                token:
                  type: string
                newPassword:
                  type: string
              required:
                - email
                - token
                - newPassword
      responses:
        "204":
          description: The password has been reset successfully.
        "400": *a1
        "401": *a1
        "403": *a1
        "404": *a1
        "500": *a1
  /auth/accept-invite:
    post:
      description: Accept an invite and set the account's password.
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                token:
                  type: string
                newPassword:
                  type: string
              required:
                - email
                - token
                - newPassword
      responses:
        "204":
          description: The password has been set successfully.
        "400": *a1
        "401": *a1
        "403": *a1
        "404": *a1
        "500": *a1
...
  /invoices:
    get:
      description: Get a list of invoices
      parameters:
        - schema:
            type: string
            enum:
              - PENDING
              - PAID
              - DECLINED
          required: false
          name: status
          in: query
        - schema:
            type: string
          required: false
          name: providerName
          in: query
        - schema:
            type: string
          required: false
          name: studentName
          in: query
        - schema:
            type: string
            enum:
              - SPEECH
              - OCCUPATIONAL
              - PHYSICAL
          required: false
          name: therapyServiceType
          in: query
        - schema:
            type:
              - string
              - "null"
          required: false
          name: dateIssuedFrom
          in: query
        - schema:
            type:
              - string
              - "null"
          required: false
          name: dateIssuedTo
          in: query
        - schema:
            type: string
            enum:
              - status
              - amount
              - dateIssued
              - createdAt
              - updatedAt
              - providerName
              - studentName
              - therapyServiceType
            default: dateIssued
          required: false
          name: sortBy
          in: query
        - schema:
            type: string
            enum:
              - asc
              - desc
            default: desc
          required: false
          name: sortOrder
          in: query
        - schema:
            type: number
            minimum: 1
            default: 20
          required: false
          name: perPage
          in: query
        - schema:
            type: number
            minimum: 1
            default: 1
          required: false
          name: page
          in: query
      responses:
        "200":
          description: Object with data and pagination metadata
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Invoice"
                  pagination:
                    $ref: "#/components/schemas/Pagination"
                required:
                  - data
                  - pagination
        "400": *a1
        "401": *a1
        "403": *a1
        "404": *a1
        "500": *a1
    post:
      description: Create an invoice
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                providerId:
                  type: string
                  format: uuid
                studentId:
                  type: string
                  format: uuid
                therapyServiceId:
                  type: string
                  format: uuid
                amount:
                  type: number
                status:
                  type: string
                  enum:
                    - PENDING
                    - PAID
                    - DECLINED
                  default: PENDING
                dateIssued:
                  type:
                    - string
                    - "null"
                  default: 2025-03-27T21:45:09.716Z
              required:
                - providerId
                - studentId
                - therapyServiceId
                - amount
      responses:
        "201":
          description: The invoice has been created successfully.
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    format: uuid
                required:
                  - id
        "400": *a1
        "401": *a1
        "403": *a1
        "404": *a1
        "500": *a1
  /invoices/{id}:
    get:
      description: Get an invoice
      parameters:
        - schema:
            type: string
            format: uuid
          required: true
          name: id
          in: path
      responses:
        "200":
          description: Object with data containing invoice information.
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/Invoice"
                required:
                  - data
        "400": *a1
        "401": *a1
        "403": *a1
        "404": *a1
        "500": *a1
    patch:
      description: Update an invoice
      parameters:
        - schema:
            type: string
            format: uuid
          required: true
          name: id
          in: path
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                providerId:
                  type: string
                  format: uuid
                studentId:
                  type: string
                  format: uuid
                therapyServiceId:
                  type: string
                  format: uuid
                amount:
                  type: number
                status:
                  type: string
                  enum:
                    - PENDING
                    - PAID
                    - DECLINED
                  default: PENDING
                dateIssued:
                  type:
                    - string
                    - "null"
                  default: 2025-03-27T21:45:09.716Z
      responses:
        "204":
          description: The invoice has been updated successfully.
        "400": *a1
        "401": *a1
        "403": *a1
        "404": *a1
        "500": *a1
    delete:
      description: Delete an invoice
      parameters:
        - schema:
            type: string
            format: uuid
          required: true
          name: id
          in: path
      responses:
        "204":
          description: The invoice has been deleted successfully.
        "400": *a1
        "401": *a1
        "403": *a1
        "404": *a1
        "500": *a1
...